Add used SDL libraries into pkgconfig file
As SDL2pp doesn't replace SDL2 and is designed to be used with plain
SDL2 code, consumers which use SDL2pp with pkgconfig will almost always
be forced to include both SDL2pp and SDL2 includes and libraries. This
seems inconvenient, so add SDL2 libraries to SDL2pp pkgconfig file. This
also improves consistency with the fact that SDL2 includes are
already included in sdl2pp.pc.