Add libmingw32 to SDL2pp library list on MINGW
Without it, SDL2pp using apps won't compile with mingw, with undefined
reference to WinMain. More investigation is needed on what happens here,
so for now just link the library. To be noted, it's the same as FindSDL
bundled with CMake does.

See also issue #66 and http://forums.libsdl.org/viewtopic.php?p=49618